+++ test_site/app/src/components/mod.rs
@@ -1,3 +1,5 @@
 mod sidebar;
+mod theme_switcher;
 
 pub use sidebar::*;
+pub use theme_switcher::*;
